revert pom.xml + some modification
