+++ src/main/resources/config.yml
@@ -66,6 +66,10 @@ enable-areashop-integration: true
 # or sell something from/to his shop or when his shop is out of stock.
 enable-vendor-messages: true
 
+# Set whether the vendor of a shop should get messages on all servers when players
+# buy or sell something from/to his shop or when his shop is out of stock.
+enable-vendor-bungee-messages: false
+
 # Set whether only the shop a player is pointing at should be shown.
 # If set to false, every shop near the player (with the specified
 # distance) will be shown to him.
